fix: make toasts visible on modal screens
